fix column size reporting

+++ index-common/src/main/java/com/metamx/druid/index/serde/DictionaryEncodedColumnPartSerde.java
@@ -49,7 +49,7 @@ public class DictionaryEncodedColumnPartSerde implements ColumnPartSerde
   private final VSizeIndexed multiValuedColumn;
   private final GenericIndexed<ImmutableConciseSet> bitmaps;
 
-  private final int size;
+  private final long size;
 
   public DictionaryEncodedColumnPartSerde(
       GenericIndexed<String> dictionary,
@@ -63,7 +63,7 @@ public class DictionaryEncodedColumnPartSerde implements ColumnPartSerde
     this.multiValuedColumn = multiValCol;
     this.bitmaps = bitmaps;
 
-    int size = dictionary.getSerializedSize();
+    long size = dictionary.getSerializedSize();
     if (singleValCol != null && multiValCol == null) {
       size += singleValCol.getSerializedSize();
     }
@@ -94,7 +94,7 @@ public class DictionaryEncodedColumnPartSerde implements ColumnPartSerde
   }
 
   @Override
-  public int numBytes()
+  public long numBytes()
   {
     return 1 + size;
   }
