+++ index-common/src/main/java/com/metamx/druid/index/serde/LongGenericColumnPartSerde.java
@@ -58,7 +58,7 @@ public class LongGenericColumnPartSerde implements ColumnPartSerde
   }
 
   @Override
-  public int numBytes()
+  public long numBytes()
   {
     return compressedLongs.getSerializedSize();
   }
